pyproject.toml
[tool.poetry]
name = "gmo-websocket"
version = "0.1.0"
description = ""
authors = ["tsugumi-sys <tidemark0105@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
asyncio = "^3.4.3"
websockets = "^10.1"
python-dotenv = "^0.19.2"
pydantic = "^1.9.0"
SQLAlchemy = "^1.4.31"
matplotlib = "^3.5.1"
requests = "^2.27.1"
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
plotly = "^5.6.0"
pandas = "^1.4.1"
numba = "^0.55.1"
seaborn = "^0.11.2"
memory-profiler = "^0.60.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"
jupyterlab = "^3.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
